# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/parties:
    get:
      tags:
        - Parties
      summary: List Parties
      description: Get a paginated list of Parties with optional filtering and sorting
      operationId: listParties
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Parties
      summary: Create Party
      description: Create a new Party
      operationId: createParty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Party'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/party_addresses:
    get:
      tags:
        - PartyAddresses
      summary: List PartyAddresses
      description: Get a paginated list of PartyAddresses with optional filtering and sorting
      operationId: listPartyAddresses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyAddressList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PartyAddresses
      summary: Create PartyAddress
      description: Create a new PartyAddress
      operationId: createPartyAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartyAddressRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyAddress'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Party:
      type: object
      required:
        - id
        - company_id
        - party_code
        - party_kind
        - name
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        party_code:
          type: string
          maxLength: 30
        party_kind:
          type: string
        name:
          type: string
          maxLength: 200
        legal_name:
          nullable: true
          type: string
          maxLength: 200
        first_name:
          nullable: true
          type: string
          maxLength: 100
        last_name:
          nullable: true
          type: string
          maxLength: 100
        npwp:
          nullable: true
          type: string
          maxLength: 25
        nik:
          nullable: true
          type: string
          maxLength: 20
        vat:
          nullable: true
          type: string
          maxLength: 32
        status:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePartyRequest:
      type: object
      required:
        - company_id
        - party_code
        - party_kind
        - name
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        party_code:
          type: string
          maxLength: 30
        party_kind:
          type: string
        name:
          type: string
          maxLength: 200
        legal_name:
          nullable: true
          type: string
          maxLength: 200
        first_name:
          nullable: true
          type: string
          maxLength: 100
        last_name:
          nullable: true
          type: string
          maxLength: 100
        npwp:
          nullable: true
          type: string
          maxLength: 25
        nik:
          nullable: true
          type: string
          maxLength: 20
        vat:
          nullable: true
          type: string
          maxLength: 32
        status:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PartyList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Party'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PartyAddress:
      type: object
      required:
        - id
        - party_id
        - company_id
        - address_type
        - line1
        - is_primary
        - is_billing
        - is_shipping
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        address_type:
          type: string
        label:
          nullable: true
          type: string
          maxLength: 60
        line1:
          type: string
          maxLength: 255
        line2:
          nullable: true
          type: string
          maxLength: 255
        country_id:
          nullable: true
          type: string
          format: uuid
        province_id:
          nullable: true
          type: string
          format: uuid
        city_id:
          nullable: true
          type: string
          format: uuid
        district_id:
          nullable: true
          type: string
          format: uuid
        subdistrict_id:
          nullable: true
          type: string
          format: uuid
        postal_code:
          nullable: true
          type: string
          maxLength: 10
        latitude:
          nullable: true
          type: string
        longitude:
          nullable: true
          type: string
        is_primary:
          type: boolean
        is_billing:
          type: boolean
        is_shipping:
          type: boolean
        status:
          type: string
        metadata:
          type: object
...
    CreatePartyAddressRequest:
      type: object
      required:
        - party_id
        - company_id
        - address_type
        - line1
        - is_primary
        - is_billing
        - is_shipping
        - status
        - metadata
      properties:
        party_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        address_type:
          type: string
        label:
          nullable: true
          type: string
          maxLength: 60
        line1:
          type: string
          maxLength: 255
        line2:
          nullable: true
          type: string
          maxLength: 255
        country_id:
          nullable: true
          type: string
          format: uuid
        province_id:
          nullable: true
          type: string
          format: uuid
        city_id:
          nullable: true
          type: string
          format: uuid
        district_id:
          nullable: true
          type: string
          format: uuid
        subdistrict_id:
          nullable: true
          type: string
          format: uuid
        postal_code:
          nullable: true
          type: string
          maxLength: 10
        latitude:
          nullable: true
          type: string
        longitude:
          nullable: true
          type: string
        is_primary:
          type: boolean
        is_billing:
          type: boolean
        is_shipping:
          type: boolean
        status:
          type: string
        metadata:
          type: object
...
    PartyAddressList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartyAddress'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []